tests/wpt/web-platform-tests/workers/modules/dedicated-worker-options-credentials.html

<!DOCTYPE html>
<title>DedicatedWorker: WorkerOptions 'credentials'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

// Determines the expected cookie value to be reported by a dedicated worker
// based on the given option. The worker reports an empty string as the actual
// cookie value if the cookie wasn't sent to the server. Otherwise, it's the
// value set by the headers file:
// "dedicated-worker-options-credentials.html.headers"
function DetermineExpectedCookieValue(options) {
  // Classic script loading should always send credentials regardless of the
  // 'credentials' option because the spec says the option takes effect only
  // for module script loading.
  if (options.type == 'classic')
    return 'COOKIE_VALUE';
  assert_equals(options.type, 'module');

  if (!options.credentials ||
      options.credentials == 'same-origin' ||
      options.credentials == 'include') {
    return 'COOKIE_VALUE';
  }
  if (options.credentials == 'omit')
    return '';
  assert_unreached('Invalid credentials option was specified: ' +
                   options.credentials);
}

// Runs a credentials test with the given WorkerOptions.
function credentials_test(options, description) {
  promise_test(async () => {
    const worker = new Worker('resources/credentials.py', options);

    // Wait until the worker sends the actual cookie value.
    const msg_event = await new Promise(resolve => worker.onmessage = resolve);

    const expectedCookieValue = DetermineExpectedCookieValue(options);
    assert_equals(msg_event.data, expectedCookieValue);
  }, description);
}

// Tests for module scripts.

credentials_test(
    { type: 'module'},
    'new Worker() with the default credentials option should behave as ' +
    'credentials=same-origin and send the credentials');

credentials_test(
    { credentials: 'omit', type: 'module' },
    'new Worker() with credentials=omit should not send the credentials');

credentials_test(
    { credentials: 'same-origin', type: 'module' },
    'new Worker() with credentials=same-origin should send the credentials');

credentials_test(
    { credentials: 'include', type: 'module' },
    'new Worker() with credentials=include should send the credentials');

// Tests for classic scripts.

credentials_test(
    { type: 'classic' },
    'new Worker() with type=classic should always send the credentials ' +
    'regardless of the credentials option (default).');

credentials_test(
    { credentials: 'omit', type: 'classic' },
    'new Worker() with type=classic should always send the credentials ' +
    'regardless of the credentials option (omit).');

credentials_test(
    { credentials: 'same-origin', type: 'classic' },
    'new Worker() with type=classic should always send the credentials ' +
    'regardless of the credentials option (same-origin).');

credentials_test(
    { credentials: 'include', type: 'classic' },
    'new Worker() with type=classic should always send the credentials ' +
    'regardless of the credentials option (include).');

</script>
